requirements.txt
cryptography>=36.0.2
opencv-python>=4.5.5.64
pysftp>=0.2.9
Pillow>=9.0.1
pynput>=1.7.6
pyinstaller>=5.0.1
